+++ tests/test_miscs.py
@@ -10,6 +10,7 @@ modulepath = os.path.dirname(os.path.dirname(thisfile))
 
 sys.path.insert(0, modulepath)
 import tensorcircuit as tc
+from tensorcircuit import experimental
 from tensorcircuit.quantum import PauliString2COO, PauliStringSum2COO
 from tensorcircuit.applications.vqes import construct_matrix_v2
 
@@ -72,8 +73,6 @@ def test_adaptive_vmap(backend):
 
     x = tc.backend.ones([30, 2])
 
-    from tensorcircuit import experimental
-
     vf = experimental.adaptive_vmap(f, chunk_size=6)
     np.testing.assert_allclose(vf(x), tc.backend.ones([30, 2]), atol=1e-5)
 
@@ -95,11 +94,47 @@ def test_adaptive_vmap_mul_io(backend):
     def f(x, y, a):
         return x + y + a
 
-    from tensorcircuit import experimental
-
     vf = experimental.adaptive_vmap(f, chunk_size=6, vectorized_argnums=(0, 1))
     x = tc.backend.ones([30, 2])
     a = tc.backend.ones([2])
     # jax vmap has some weird behavior in terms of keyword arguments...
     # TODO(@refraction-ray): further investigate jax vmap behavior with kwargs
     np.testing.assert_allclose(vf(x, x, a), 3 * tc.backend.ones([30, 2]), atol=1e-5)
+
+
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_qng(backend):
+    n = 6
+
+    def f(params):
+        params = tc.backend.reshape(params, [4, n])
+        c = tc.Circuit(n)
+        c = tc.templates.blocks.example_block(c, params)
+        return c.state()
+
+    params = tc.backend.ones([4 * n])
+    fim = experimental.qng(f)(params)
+    assert tc.backend.shape_tuple(fim) == (4 * n, 4 * n)
+    print(experimental.dynamics_matrix(f)(params))
+
+
+@pytest.mark.parametrize("backend", [lf("tfb"), lf("jaxb")])
+def test_dynamic_rhs(backend):
+    h1 = tc.array_to_tensor(tc.gates._z_matrix)
+
+    def f(param):
+        c = tc.Circuit(1)
+        c.rx(0, theta=param)
+        return c.state()
+
+    rhsf = experimental.dynamics_rhs(f, h1)
+    np.testing.assert_allclose(rhsf(tc.backend.ones([])), -np.sin(1.0) / 2, atol=1e-5)
+
+    h2 = tc.backend.coo_sparse_matrix(
+        indices=tc.array_to_tensor(np.array([[0, 0], [1, 1]]), dtype="int64"),
+        values=tc.array_to_tensor(np.array([1, -1])),
+        shape=[2, 2],
+    )
+
+    rhsf = experimental.dynamics_rhs(f, h2)
+    np.testing.assert_allclose(rhsf(tc.backend.ones([])), -np.sin(1.0) / 2, atol=1e-5)